fix: Do not directly update state during ngAfterViewInit
Directly manipulating component state at this point in the lifecycle is not recommended since it occurs after change detection. Whether or not this fixes the errors we're seeing will need to be evaluated on the server but this change is still more in-line with how updates should be handled in that phase.

Also includes a refactor that moves the scorm API implementation into a class of its own. The result is a much cleaner component that makes debugging easier shoudl more be needed.