feat: More thorough null check
Adds a new utility function called `isNullObject` to the `CourseEnrollmentComponent`. This function checks if an object is null or undefined, or if its `id` property is 0. It is used in the `filter` operator to ensure that both the course and order objects are not null before proceeding with further operations.